Refactor Python API tests and added natural lanuage output
- Consolidate 14 seperate test files into 8 logica modules
- Create shared helpers.py to create required dummy shared resources
- Updat README with  new test files description and usage
