+++ entity/python/tests/test_keys.py
@@ -3,20 +3,11 @@ import unittest
 from iotauth import KeyCacheError, SessionKey, SessionKeyCache
 
 
-def make_key(key_id=b"12345678"):
-    return SessionKey(
-        id=key_id,
-        cipher_key=b"c" * 16,
-        mac_key=b"m" * 32,
-        abs_validity=None,
-        rel_validity=None,
-        encryption_mode="AES_128_CBC",
-        hmac_enabled=True,
-        permanent_distribution_key=False,
-    )
-
+from tests.helpers import make_session_key
 
 class SessionKeyCacheTests(unittest.TestCase):
+    """Tests for the session key cache data structure."""
+
     def test_cache_starts_empty(self):
         cache = SessionKeyCache()
 
@@ -25,7 +16,7 @@ class SessionKeyCacheTests(unittest.TestCase):
 
     def test_adds_and_retrieves_key_by_id(self):
         cache = SessionKeyCache()
-        key = make_key()
+        key = make_session_key()
 
         cache.add(key)
 
@@ -35,19 +26,19 @@ class SessionKeyCacheTests(unittest.TestCase):
 
     def test_rejects_invalid_key_id_size(self):
         with self.assertRaisesRegex(KeyCacheError, "8 bytes"):
-            make_key(b"short")
+            make_session_key(b"short")
 
     def test_rejects_duplicate_key_without_replace(self):
         cache = SessionKeyCache()
-        cache.add(make_key())
+        cache.add(make_session_key())
 
         with self.assertRaisesRegex(KeyCacheError, "already exists"):
-            cache.add(make_key())
+            cache.add(make_session_key())
 
     def test_replace_allows_existing_key_update(self):
         cache = SessionKeyCache()
-        old_key = make_key()
-        new_key = make_key()
+        old_key = make_session_key()
+        new_key = make_session_key()
         cache.add(old_key)
 
         cache.add(new_key, replace=True)
@@ -56,11 +47,11 @@ class SessionKeyCacheTests(unittest.TestCase):
 
     def test_enforces_max_key_count(self):
         cache = SessionKeyCache(max_keys=1)
-        cache.add(make_key())
+        cache.add(make_session_key())
 
         with self.assertRaisesRegex(KeyCacheError, "full"):
-            cache.add(make_key(b"87654321"))
+            cache.add(make_session_key(b"87654321"))
 
 
 if __name__ == "__main__":
-    unittest.main()
+    unittest.main(verbosity=2)
